requirements.txt
# 06-cloud-soc-canada — Requirements
# Cloud-Native SOC Monitoring (AWS & Azure — Canada Region)
# pip install -r requirements.txt

# AWS SDK
boto3==1.34.69
botocore==1.34.69

# Azure SDKs
azure-identity==1.16.0
azure-monitor-query==1.3.0
azure-mgmt-security==6.0.0
azure-mgmt-compute==30.6.0
azure-mgmt-network==25.3.0
azure-mgmt-resource==23.0.1
azure-mgmt-monitor==6.0.1

# HTTP & API
requests==2.31.0
urllib3==2.2.1

# Environment & Config
python-dotenv==1.0.1

# Terminal Output & Logging
rich==13.7.1
colorama==0.4.6
tqdm==4.66.2

# Data Processing
pandas==2.2.1
tabulate==0.9.0

# Elasticsearch Integration
elasticsearch==8.13.0

# Task Scheduling
schedule==1.2.1

# Testing
pytest==8.1.1
pytest-cov==5.0.0
pytest-mock==3.12.0

# JWT / Auth utilities
python-jose==3.3.0

# Compression (stdlib supplement)
zstandard==0.22.0
